life_pulse_backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
pydantic>=2.10.0
apscheduler>=3.10.4

# Testing
pytest>=8.3.3
httpx>=0.27.2
